Expand mypy to whole project and fix all type errors
CI/local mypy was limited to quant_rl (with a large exclude list), so
mt5_trading, scripts, tests, backtest and the root demo/live scripts
were never type-checked. pyproject.toml now lists every package plus the
root scripts under [tool.mypy], and the exclude regex is gone.

Fixed the ~477 errors this surfaced:
- quant_rl: lazy cfg/Any typing in make_env, typed callbacks list,
  CheckpointCallback.save_path=str(model_dir), wandb added to
  ignore_missing_imports, scipy stubs, pandas/Figure casts in plots
- scripts: report_g3/discover_symbols/train_vae/run_baseline_eval fully
  annotated (dict[str, Any], Dataset[T], DataLoader[T], cast helpers)
- mt5_trading: strategy ABC uses TYPE_CHECKING for Signal (breaks the
  runtime chain into MetaTrader5, which is Windows-only); robots and
  manager got return types and dict[str, Any] annotations
- tests: dict/ndarray/tuple type args, get_indexer(pd.Index([...])),
  cast(pd.Timestamp, ...) for Hashable indices, protocol-typed stubs
- root scripts: demo_trading/live_trading null-safe credential guards

CI: ci-scripts/mypy and docker-compose typecheck service now run plain
mypy (project-wide via pyproject.toml) instead of mypy quant_rl.

Final: mypy 0 errors in 163 files, ruff clean, 351 tests pass.

+++ tests/test_eval/test_evaluate_metrics_contract.py
@@ -114,7 +114,7 @@ def test_evaluate_model_output_satisfies_metrics_contract() -> None:
         "evaluate_model must return a pd.Series for equity "
         "(calculate_metrics relies on .to_numpy())"
     )
-    eq = result["equity"].to_numpy(dtype=float)
+    eq = result["equity"].to_numpy(dtype=np.float32)
     assert eq.size > 0
     assert eq[0] > 0, "initial equity must be positive (calculate_metrics rejects <= 0)"
